improve styling

+++ client/src/components/ReportEdit.js
@@ -10,7 +10,6 @@ const ReportEdit = () => {
   let params = useParams()
   const location = useLocation()
   const navigate = useNavigate()
-  const id = params.id
 
   useEffect(() => {
     if (location.state.report) {
@@ -18,6 +17,8 @@ const ReportEdit = () => {
       setRepCount(location.state.report.rep_count);
     }
   }, [location.state.report]);
+
+  const id = params.id
 
   async function handleSubmit(e) {
     e.preventDefault()
@@ -36,38 +37,42 @@ const ReportEdit = () => {
 
   async function deleteReport(e) {
     e.preventDefault()
-    let data = await axios.delete(`${API_ROOT}/api/reports/${id}`)
-    if (data) {
-      alert('report deleted!')
-      navigate("/");
+    if (window.confirm("Are you certain you want to delete your report?")) {
+      let data = await axios.delete(`${API_ROOT}/api/reports/${id}`)
+      if (data) {
+        alert('report deleted!')
+        navigate("/");
+      }
     }
   }
 
   if (location.state){
     return (
-      <div>
-      <Form onSubmit={e => handleSubmit(e)}>
-        <Form.Group className="mb-1" controlId="rep_type">
-          <Form.Control 
-            type="text" 
-            placeholder={location.state.report.rep_type}
-            name="rep_type"
-            value={rep_type}
-            onChange={e => setRepType(e.target.value)} 
-          />
-        </Form.Group>
-        <Form.Group className="mb-1" controlId="rep_count">
-          <Form.Control 
-            type="number" 
-            placeholder={location.state.report.rep_count} 
-            name="rep_count"
-            value={rep_count} 
-            onChange={e => setRepCount(e.target.value)} 
-          />
-        </Form.Group>
-      <Button variant="primary" type="submit">Update</Button>
-    </Form>
-    <Button variant="primary" onClick={deleteReport}>Delete</Button>
+      <div className="container">
+        <Form className="form-field" onSubmit={e => handleSubmit(e)}>
+          <Form.Group className="mb-1" controlId="rep_type">
+            <Form.Control 
+              type="text" 
+              placeholder={location.state.report.rep_type}
+              name="rep_type"
+              value={rep_type}
+              onChange={e => setRepType(e.target.value)} 
+            />
+          </Form.Group>
+          <Form.Group className="mb-1" controlId="rep_count">
+            <Form.Control 
+              type="number" 
+              placeholder={location.state.report.rep_count} 
+              name="rep_count"
+              value={rep_count} 
+              onChange={e => setRepCount(e.target.value)} 
+            />
+          </Form.Group>
+        <div>
+        <Button size="sm" className="button" variant="primary" type="submit">Update</Button>
+        <Button size="sm" className="button" variant="primary" onClick={deleteReport}>Delete</Button>
+        </div>
+      </Form>
     </div>
     )
   } else {
